requirement.txt
aioredis==1.3.1
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Brlapi==0.6.6
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==2.9.2
cupshelpers==1.0
daphne==2.5.0
defer==1.0.6
defusedxml==0.7.0rc1
distro-info==0.18
Django==3.0.8
django-allauth==0.42.0
django-braces==1.14.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-star-ratings==0.9.1
docutils==0.14
hiredis==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
keyring==10.6.0
keyrings.alt==3.0
Kivy==1.11.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
msgpack==0.6.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pdfminer.six==20200517
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pyRFC3339==1.0
python-apt==1.6.2
python-debian==0.1.32
python-decouple==3.3
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
roman==2.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.13.2
six==1.15.0
social-auth-core==3.3.3
sortedcontainers==2.2.2
sqlparse==0.3.1
ssh-import-id==5.7
swapper==1.1.2.post1
system-service==0.3
systemd-python==234
Twisted==20.3.0
txaio==20.4.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
wadllib==1.3.2
xkit==0.0.0
zope.interface==5.1.0
